fox temp

+++ Driver/Tank.pde
@@ -115,7 +115,7 @@
             setPos(x, y);
         }
         public void detection(float xDist, float yDist) {
-            int r=Math.sqrt(xDist*xDist+yDist*yDist);
+            int r=(int)Math.sqrt(xDist*xDist+yDist*yDist);
             if (r<= 10) {
                 setDetection(true);
             }
